데미지, 체력 타입을 float으로 변경

+++ Valheim/Grass.cpp
@@ -35,7 +35,7 @@ void Grass::Release()
 {
 }
 
-bool Grass::ReceivedDamageEvent(int damage)
+bool Grass::ReceivedDamageEvent(float damage)
 {
 	return false;
 }
